gen3-minio-client/requirements.txt
aiofiles==0.8.0
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncio==3.4.3
attrs==23.2.0
backoff==1.11.1
boto3==1.34.125
botocore==1.34.125
cdislogging==1.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.9
dictionaryutils==3.4.10
drsclient==0.2.3
fastavro==1.8.4
frozenlist==1.4.1
gen3==4.24.2
gen3Dictionary==2.0.3
gen3users==1.0.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
idna==3.7
indexclient==2.3.1
jmespath==1.0.1
jsonschema==3.2.0
marshmallow==3.21.3
marshmallow-enum==1.5.1
minio==7.2.7
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pycparser==2.22
pycryptodome==3.20.0
pypfb==0.5.29
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==0.1.11
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rfc3986==1.5.0
s3transfer==0.10.1
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
xmltodict==0.13.0
yarl==1.9.4
